+++ lib/src/output_format.dart
@@ -7,5 +7,4 @@
 /// .flv,
 /// .avi,
 /// .wmv,
-/// .gif.
-enum OutputFormat { mp4, mkv, mov, flv, avi, wmv, gif }
+enum OutputFormat { mp4, mkv, mov, flv, avi, wmv }