+++ lib/src/watermark_source.dart
@@ -1,43 +1,58 @@
 import 'dart:io';
-import 'package:flutter/services.dart' show ByteData, rootBundle;
+import 'package:flutter/services.dart'
+    show ByteData, NetworkAssetBundle, rootBundle;
 import 'package:path_provider/path_provider.dart'
     show getApplicationDocumentsDirectory;
 
+enum WatermarkSourceEnum { file, asset, network }
+
 /// Add images from various sources in video as overlay using [WatermarkSource].
 class WatermarkSource {
-  String? _path;
+  String _path = '';
+  WatermarkSourceEnum? _source;
 
   /// To add image from local storage with file's path
   WatermarkSource.file(String path) {
     _path = path;
+    _source = WatermarkSourceEnum.file;
   }
 
   /// To add image from assets with asset's path.
   WatermarkSource.asset(String path) {
-    ByteData byteData;
-    String appDocDir;
-    File file;
-    rootBundle.load(path).then((value) {
-      byteData = value;
-      getApplicationDocumentsDirectory().then((value) {
-        appDocDir = value.path;
-        File('$appDocDir/$path').create(recursive: true).then((value) {
-          file = value;
-          file
-              .writeAsBytes(byteData.buffer
-                  .asUint8List(byteData.offsetInBytes, byteData.lengthInBytes))
-              .then(
-            (value) {
-              _path = file.path;
-            },
-          );
-        });
-      });
-    });
+    _path = path;
+    _source = WatermarkSourceEnum.asset;
   }
 
-  @override
-  String toString() {
-    return _path ?? "null";
+  /// To add image from network with image url.
+  WatermarkSource.network(String path) {
+    _path = path;
+    _source = WatermarkSourceEnum.network;
+  }
+
+  Future<String> toCommand() async {
+    if (_path.isNotEmpty) {
+      try {
+        if (_source != WatermarkSourceEnum.file) {
+          ByteData byteData;
+          String appDocDir;
+          File file;
+          byteData = _source == WatermarkSourceEnum.asset
+              ? await rootBundle.load(_path)
+              : await NetworkAssetBundle(Uri.parse(_path)).load("");
+          appDocDir = await getApplicationDocumentsDirectory()
+              .then((value) => value.path);
+          file = await File('$appDocDir/$_path').create(recursive: true);
+          _path = await file
+              .writeAsBytes(byteData.buffer
+                  .asUint8List(byteData.offsetInBytes, byteData.lengthInBytes))
+              .then((value) => value.path);
+        }
+      } catch (e) {
+        throw ("Unable to fetch image from the mentioned source.");
+      }
+      return _path;
+    } else {
+      throw ("Add Image source using WatermarkSource");
+    }
   }
 }