+++ CHANGELOG.md
@@ -5,3 +5,10 @@
 # 1.0.1
 
 * Support to add images from File and assets.
+
+# 1.0.2
+
+* Added new watermark image source - `Network`.
+* Fixed bug in asset watermark source.
+* Temporarly not supporting output format `.gif`
+* Added support from Android SDK 16 and IOS SDK 10.